Acrhitecture factorisation
Modified the structure of the repository.
Added new services, each focused on one specific: frontend, server for frontend, pyannote api wrapper, mongo gateway for every modifications on the collections.
Modified containers communication adding private network. The goal is to only expose the port of the frontend's container.
Right now, the user can upload a file on PyAnnote storage file system (for 48 hours). It also registers the file in a local database.