Webhook processing the results of the diarization job submitted
Added endpoint in the webhook container which validates the origin of the post request and processes the result.
Then the container sends to the main server the results. As the main server gets the request he asks to the mongo gateway to update the infos of the file whose diarization is finished.